Change liftIO to have an unrestricted arrow type
Since it's currently not possible to map linear functions in arbitrary
functors, it's impossible to implement HasIO for monad transformers.
This makes it useless as a MTL-style effects interface.

I don't think that getting rid of the current linearity requirement
would break any code, but it would allow for HasIO to be used in many
more situations.

(OTOH if multiplicity polymorphism ever does come to fmap, linear liftIO
would be great.)